moved @type deps to devDependencies
Hey!
We're using this dependency in one of our projects, and we came to realize that once we install it our types break. This is because we use lodash, and once installing yes-kafka it also installs type definitions for lodash as well, which break our typescript compiler.

In order to circumvent that, I've moved all @type deps to devDependencies. Type definitions should not be under regular deps for that exact reason - they can clash with the host application's configuration and type dependencies :)